+++ packages/@leight/file-client/src/sdk/ClientSourceProvider.tsx
@@ -59,4 +59,4 @@ export const FileQueryProvider:FC<IFileQueryProviderProps> = props => {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_rbck6up0u6w67skx9hlfw2cu = true;
+export const $leight_nxvmysnghy35dwzjjzicstus = true;