+++ packages/@leight/file-client/src/sdk/ClientStore.ts
@@ -43,4 +43,4 @@ export const FileSortStore = createSortContext<IFileSortSchema>({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_aflw8yjys7ghvoo9oprdf3ry = true;
+export const $leight_yvo3yt2t34sg55ivle6t02ox = true;