+++ packages/@leight/file-server/src/sdk/ServerTrpc.ts
@@ -18,4 +18,4 @@ export const FileSourceProcedure = withSourceProcedure<IFileSourceSchema>({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_xocja9dgltaow7qt9nora5jx = true;
+export const $leight_j7yf7sysinkgev2hcm0voo9w = true;