+++ packages/@leight/file/src/sdk/PrismaSchema.ts
@@ -71,4 +71,4 @@ export const FileQuerySchema = QuerySchema({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_ilad8rrt3t78ar0mlnrjqz2i = true;
+export const $leight_g7rno3lce34sz60bxv0r0eki = true;