+++ packages/@leight/file/src/sdk/Source.ts
@@ -19,4 +19,4 @@ export const FileSourceContext = (container: IContainer) => new ServiceContext<I
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_cvzipxnvu9y5a0ck5o3ha16d = true;
+export const $leight_wiutocb2a2hal2yadq7w8hku = true;