+++ packages/@leight/file/src/sdk/SourceSchema.ts
@@ -34,4 +34,4 @@ export interface IFileSourceSchema extends ISourceSchema<
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_n90yxwzdtxvpaob16yf624jx = true;
+export const $leight_waqflzv18uy16jy2lkgftci1 = true;