+++ packages/@leight/job-client/src/sdk/ClientSourceProvider.tsx
@@ -59,4 +59,4 @@ export const JobQueryProvider:FC<IJobQueryProviderProps> = props => {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_oxozshud8cf4l6qogo1y0vvp = true;
+export const $leight_rl4fsykmkhoebnqzet86mgfk = true;