+++ packages/@leight/job-client/src/sdk/ClientStore.ts
@@ -43,4 +43,4 @@ export const JobSortStore = createSortContext<IJobSortSchema>({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_r19jtnihalfjg8kqvu220g5n = true;
+export const $leight_sedtuscq6g5snnyxykypg64j = true;