+++ packages/@leight/job-server/src/sdk/ServerTrpc.ts
@@ -18,4 +18,4 @@ export const JobSourceProcedure = withSourceProcedure<IJobSourceSchema>({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_wzjys9u5e0s483ri2h39z94u = true;
+export const $leight_fpk696yn1tjhl0dnaa0h74sn = true;