+++ packages/@leight/job/src/sdk/PrismaSchema.ts
@@ -72,4 +72,4 @@ export const JobQuerySchema = QuerySchema({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_karalchg1djp1sygmkgadsjr = true;
+export const $leight_u9ui07uajgvhzi3utxgd398y = true;