+++ packages/@leight/job/src/sdk/Source.ts
@@ -19,4 +19,4 @@ export const JobSourceContext = (container: IContainer) => new ServiceContext<IJ
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_ofiqw6i829d89zw2gu89c0su = true;
+export const $leight_p4i4283x8ml796o7a2189wft = true;