+++ packages/@leight/job/src/sdk/SourceSchema.ts
@@ -34,4 +34,4 @@ export interface IJobSourceSchema extends ISourceSchema<
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_k64mt2xzj0zp687ha0ijjk5h = true;
+export const $leight_g34ex0m5meprzv12xulcd2rh = true;