+++ packages/@leight/mantine/src/component/index.ts
@@ -1,4 +1,5 @@
 export * from "./DropZone";
+export * from "./MainMenu";
 export * from "./PageShell";
 export * from "./Paper";
 export * from "./Shell";