+++ packages/@leight/user/src/sdk/Source.ts
@@ -20,4 +20,4 @@ export const UserSourceContext = (container: IContainer) => new ServiceContext<I
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_we7zsdgk2nljy4ju0hua1nme = true;
+export const $leight_k0yrutho2q45k8kkmbvn34ov = true;