+++ packages/@leight/user/src/sdk/SourceSchema.ts
@@ -34,4 +34,4 @@ export interface IUserSourceSchema extends ISourceSchema<
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_r8fr2j8fpox5yg2dufr7crdv = true;
+export const $leight_ipubfesu4cnkbcha4dvybi1h = true;