+++ packages/@leight/container/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/container
 
+## 0.5.64
+
+### Patch Changes
+
+- Version bump (updated packages)
+
 ## 0.5.63
 
 ### Patch Changes