+++ packages/@leight/esbuild/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/esbuild
 
+## 0.5.42
+
+### Patch Changes
+
+- Version bump (updated packages)
+
 ## 0.5.41
 
 ### Patch Changes