+++ packages/@leight/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/utils
 
+## 0.5.66
+
+### Patch Changes
+
+- Version bump (updated packages)
+
 ## 0.5.65
 
 ### Patch Changes