+++ packages/@leight/viv-client/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @leight/viv-client
 
+## 0.5.2
+
+### Patch Changes
+
+- Version bump (updated packages)
+- Updated dependencies
+  - @leight/mantine@0.5.70
+  - @leight/calendar-client@0.5.21
+  - @leight/context-client@0.5.71
+
 ## 0.5.1
 
 ### Patch Changes