+++ packages/@leight/winston/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/winston
 
+## 0.5.62
+
+### Patch Changes
+
+- Version bump (updated packages)
+
 ## 0.5.61
 
 ### Patch Changes